Code cleanup

+++ src/Board.h
@@ -135,9 +135,9 @@ extern S_HASHTABLE HashTable[1];
 extern Bitboard SQUARES_BETWEEN_BB[Board_sq_num][Board_sq_num];
 
 typedef struct info {
-	int starttime = -1;
-	int stoptimeOpt = -1;
-	int stoptimeMax = -1;
+	int64_t starttime = -1;
+	int64_t stoptimeOpt = -1;
+	int64_t stoptimeMax = -1;
 	int depth = -1;
 	int seldepth = -1;
 	bool depthset = false;
